% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_dataset}
\alias{download_dataset}
\title{Download one of the datasets from TumourMethData}
\usage{
download_dataset(dataset, dir = ".")
}
\arguments{
\item{dataset}{Name of the dataset to download. Must be one of the datsets listed in data(TumourMethDatasets).}

\item{dir}{Parent directory to save the HDF5 SummarizedExperiment datasets. A directory with the dataset name
will be created within this directory and populated with the RDS and HDF5 files constituting the dataset.}
}
\value{
A RangedSummarizedExperiment with methylation values from the specified dataset.
}
\description{
Download one of the datasets from TumourMethData
}
\examples{
tcga_wgbs_hg38 = TumourMethData::download_dataset(dataset = "tcga_wgbs_hg38", dir = ".")
print(rse)
}
