% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_meth_dataset}
\alias{download_meth_dataset}
\title{Download a WGBS methylation dataset from TumourMethData}
\usage{
download_meth_dataset(dataset, dir = getwd())
}
\arguments{
\item{dataset}{Name of the dataset to download WGBS data from. Must be one of the datsets listed in data(TumourMethDatasets).}

\item{dir}{Directory in which to save a the copy of the RDS file pointing to the
location of the HDF5 file in the ExperimentHub cache. Default is the current directory.}
}
\value{
A RangedSummarizedExperiment with methylation values from the specified dataset.
}
\description{
The HDF5 file and RDS file to construct a RangedSummarizedExperiment for the specified dataset
are downloaded into the ExperimentHub cache located at \code{ExperimentHub::getExperimentHubOption("cache")}.
A copy of the RDS file pointing to the HDF5 file in the cache is saved in the specified directory.
}
\examples{
mcrpc_wgbs_hg38_chr11 = TumourMethData::download_meth_dataset(dataset = "mcrpc_wgbs_hg38_chr11")
print(mcrpc_wgbs_hg38_chr11)
}
