% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_rnaseq_dataset}
\alias{download_rnaseq_dataset}
\title{Download a RNA-Seq counts dataset from TumourMethData}
\usage{
download_rnaseq_dataset(dataset)
}
\arguments{
\item{dataset}{Name of the dataset to download. Must be one of the datasets
listed in data(TumourMethDatasets) where \code{transcript_counts_available} is TRUE.}
}
\value{
A data.frame with RNA-Seq counts calcualted using Kallisto.
}
\description{
Download a RNA-Seq counts dataset from TumourMethData
}
\examples{
mcrpc_rnaseq = TumourMethData::download_rnaseq_dataset(dataset = "mcrpc_wgbs_hg38_chr11")
print(mcrpc_rnaseq)
}
