% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TumourMethData.R
\docType{package}
\name{TumourMethData-package}
\alias{TumourMethData}
\alias{TumourMethData-package}
\title{TumourMethData: A collection of DNA methylation datasets for human tumour samples and matching normal samples}
\description{
TumourMethData collects tumour methylation data from a variety of different tumour types
(and also matching normal samples where available) and produced with different technologies
(e.g. WGBS, RRBS and methylation arrays) and provides them as RangedSummarizedExperiments,
facilitating easy extraction of methylation data for regions of interest. At present,
includes the following datasets:
}
\details{
\itemize{
\item cpgea_wgbs_hg38 : WGBS Data (CpG sites only) from 187 pairs of matching human prostate tumours and normal prostate samples from 'Li, Jing, et al. "A genomic and epigenomic atlas of prostate cancer in Asian populations." Nature 580.7801 (2020): 93-99.'
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richardheery/TumourMethData}
  \item Report bugs at \url{https://support.bioconductor.org/tag/TumourMethData}
}

}
\author{
Richard Heery
}
