% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_rnaseq_dataset}
\alias{download_rnaseq_dataset}
\title{Download an RNA-Seq counts dataset from TumourMethData}
\usage{
download_rnaseq_dataset(dataset)
}
\arguments{
\item{dataset}{Name of the dataset to download. Must be one of the datasets
listed in data(TumourMethDatasets) where \code{transcript_counts_available} is TRUE.}
}
\value{
A data.frame with RNA-Seq counts calculated using Kallisto.
}
\description{
A TSV file with the RNA-Seq counts is downloaded into the ExperimentHub cache
located at \code{ExperimentHub::getExperimentHubOption("cache")} and is read into R as a data.frame.
}
